Updated UserRole class to use int instead of longs. Created UserWithRole class.

+++ src/main/java/com/givehopeweb/models/UserRole.java
@@ -11,27 +11,27 @@ public class UserRole {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private int id;
 
     @Column (name = "user_id")
-    private long userId;
+    private int userId;
 
     @Column (name = "role")
     private String role;
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
-    public long getUserId() {
+    public int getUserId() {
         return userId;
     }
 
-    public void setUserId(long userId) {
+    public void setUserId(int userId) {
         this.userId = userId;
     }
